requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
numpy
openai
sentence-transformers
google-generativeai
python-dotenv
